+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed each account showing the same number of matches on the Accounts page [#15](https://github.com/cheshire137/competiwatch-desktop/issues/15)
 - Fixed the blank, black tooltip that appeared when hovering over a competitive rank image
 - Fixed each database file path being the same on the Settings page [#14](https://github.com/cheshire137/competiwatch-desktop/issues/14)
+- Increased default window width to reduce horizontal scrolling on the Matches page [#13](https://github.com/cheshire137/competiwatch-desktop/issues/13)
 
 ## Version 0.1.1