+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change log
 
+## Version 0.1.3
+
 ## Version 0.1.2
 
 - Added a 'Trends' page with charts about your match results [#12](https://github.com/cheshire137/competiwatch-desktop/issues/12)